Implement professional sidebar layout with enhanced navigation and user controls
