+++ app/free-agency/page.tsx
@@ -18,7 +18,7 @@ export default function FreeAgencyPage({
   searchParams?: { [key: string]: string | string[] | undefined }
 }) {
   return (
-    <div className="min-h-screen relative overflow-hidden bg-background">
+    <div className="min-h-screen relative overflow-hidden bg-background pt-4">
       {/* Professional Free Agent Market Background */}
       <div className="absolute inset-0 overflow-hidden">
         <div className="absolute inset-0 hockey-grid opacity-20" />
